requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0
faker==18.11.2
streamlit==1.26.0